Set up comments to work on frontend for each professor page and only allow submit when logged in
