+++ web/src/components/ProfessorPage.js
@@ -4,6 +4,7 @@ import '../styles/ProfessorPageStyling.css';
 import Chart from './Chart';
 import DisplayAverageRating from './DisplayAverageRating.js';
 import SubmitRating from './SubmitRating.js';
+import Comment2 from './Comment.js';
 
 const ProfessorPage = () => {
   const { professorId } = useParams();
@@ -99,16 +100,7 @@ const ProfessorPage = () => {
       </section>
 
       <section className="reviews">
-        <h2>Student Reviews</h2>
-        <div className="review-list">
-          <p>No reviews yet. Be the first to leave one!</p>
-        </div>
-      </section>
-
-      <section className="leave-review">
-        <h3>Leave a Review</h3>
-        <textarea placeholder="Write your review here..." rows="5"></textarea>
-        <button type="submit">Submit Review</button>
+        <Comment instructorName={instructorName} />
       </section>
     </main>
   );
